requirements.txt
# Python package requirements for the ulria project.
numpy
scipy
scikit-learn
torch
torchvision
lpips
tqdm
matplotlib
einops
pyyaml
